requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.0.0
Flask>=3.0.0
PyYAML>=6.0.1
onnx>=1.15.0
onnxruntime>=1.16.0
scikit-learn>=1.3.0
pandas>=2.1.0
matplotlib>=3.8.0
numpy>=1.26.0
